+++ configs/language.config.js
@@ -1,0 +1,46 @@
+const { CPP, C, PYTHON, JAVA, NODEJS, PROMPTV1 } = require("../enums/supportedLanguages")
+const ONE_MB = 1024 // ulimit uses Kilobyte as base unit
+const ALLOWED_RAM = 512
+
+const LANGUAGES_CONFIG = {
+    [C]: {
+        compile: 'gcc -o a.out solution.c',
+        run: './a.out',
+        timeout: 2,
+        filename: 'solution.c',
+        memory: ALLOWED_RAM * ONE_MB,
+    },
+    [CPP]: {
+        compile: 'g++ -o a.out -pthread -O0 solution.cpp',
+        run: './a.out',
+        timeout: 2,
+        filename: 'solution.cpp',
+        memory: ALLOWED_RAM * ONE_MB,
+    },
+    [PYTHON]: {
+        compile: 'python -m compileall -q solution.py',
+        run: 'python solution.py',
+        timeout: 10,
+        filename: 'solution.py',
+        memory: ALLOWED_RAM * ONE_MB,
+    },
+    [JAVA]: {
+        compile: 'javac Solution.java',
+        run: 'java Solution',
+        timeout: 4,
+        filename: 'Solution.java',
+        memory: ALLOWED_RAM * ONE_MB,
+    },
+    [NODEJS]: {
+        compile: 'node --check solution.js',
+        run: 'node solution.js',
+        timeout: 10,
+        filename: 'solution.js',
+        memory: ALLOWED_RAM * ONE_MB,
+    },
+    [PROMPTV1]: {
+        model: "gpt-4-1106-preview",
+    }
+}
+
+module.exports = {LANGUAGES_CONFIG}
