+++ package.json
@@ -20,7 +20,7 @@
     "joi": "^17.7.0",
     "morgan": "^1.10.0",
     "node-cache": "^5.1.2",
-    "openai": "^4.24.1",
+    "openai": "~4.24.1",
     "uuid": "^9.0.0",
     "winston": "^3.11.0"
   },
