+++ validators/code.validator.js
@@ -1,17 +1,18 @@
 const Joi = require('joi')
+const { PROMPTV1 } = require('../enums/supportedLanguages')
 
 const isValidForExecute = async (body) => {
     const schema = Joi.object({
         language: Joi.string().required(),
         script: Joi.string()
             .when('language', {
-                is: 'promptv1',
+                is: PROMPTV1,
                 then: Joi.optional(),
                 otherwise: Joi.required()
             }),
         prompt: Joi.string()
             .when('language', {
-                is: 'promptv1',
+                is: PROMPTV1,
                 then: Joi.required(),
                 otherwise: Joi.optional()
             }),
